feat: definition and codec sync
