+++ src/codecs/domain/index.ts
@@ -1,3 +1,4 @@
 export * from './traits.codecs';
 export * from './regular.codecs';
 export * from './fares.codecs';
+export * from './destination.codec';
