+++ src/definitions/endpoints.definitions.ts
@@ -30,3 +30,11 @@ export type PendingScheduled = {
 export type RegularRegistered = {
   regularRegistered: Entity & RegularDetails;
 };
+
+export type RegularDeleted = {
+  regularDeleted: Entity & RegularDetails;
+};
+
+export type RegularEdited = {
+  regularEdited: Entity & RegularDetails;
+};
